+++ src/scripts/events/formHtml.js
@@ -1,6 +1,6 @@
 // author:Michelle Johnson (build form)
 
-import data from "../events/data.js"
+import data from "./data.js"
 import { format } from "url"
 
 
@@ -29,18 +29,8 @@ const eventCalendar = {
 
         const eventCalendarContainer = document.querySelector("#events-container")
         eventCalendarContainer.innerHTML = calendar
-    },
-    makeEventComponent: (eventEntry) => {
-        return `
-        <section>
-        <h3>${eventEntry.name}</h3>
-        <p>${eventEntry.date}</p>
-        <p>${eventEntry.location}</p>
-        <button id="editEntry--${eventEntry.id}">Edit Event</button>
-        <button id="deleteEntry--${eventEntry.id}">Delete Event</button>
-      </section>
-    `
     }
+    
     
 }
