+++ src/scripts/welcomePage/userMainPage.js
@@ -1,4 +1,4 @@
-import eventCalendar from "../events/form.js"
+import eventCalendar from "../events/formHtml.js"
 import messagesHtml from "../messages/htmlMaker"
 import messagesData from "../messages/data"
